+++ src/lib/components/overlays/overlay/Overlay.svelte
@@ -14,6 +14,8 @@
     import type {IPaddingProperties, PROPERTY_SPACING_BREAKPOINT} from "../../../types/spacings";
 
     import {auto_focus} from "../../../actions/auto_focus";
+    import {click_inside} from "../../../actions/click_inside";
+    import {click_outside} from "../../../actions/click_outside";
     import type {IForwardedActions} from "../../../actions/forward_actions";
     import {forward_actions} from "../../../actions/forward_actions";
     import {trap_focus} from "../../../actions/trap_focus";
@@ -22,6 +24,7 @@
     import {make_state_context} from "../../../stores/state";
 
     import {map_data_attributes, map_global_attributes} from "../../../util/attributes";
+    import {action_exit} from "../../../util/keybind";
 
     import ContextBackdrop from "../../utilities/contextbackdrop/ContextBackdrop.svelte";
 
@@ -101,13 +104,12 @@
         state = (event.target as HTMLInputElement).checked;
     }
 
-    function on_click_inside(event: MouseEvent): void {
-        if (once) {
-            const target = event.target as HTMLElement;
-            if (target.tagName !== "LABEL" && target.getAttribute("for") !== logic_id) {
-                state = false;
-            }
-        }
+    function on_dismiss(): void {
+        if (dismissible && $_overlay_state) state = false;
+    }
+
+    function on_once(): void {
+        if (once && $_overlay_state) state = false;
     }
 
     $: $_overlay_state = state as boolean;
@@ -120,6 +122,8 @@
         }
     }
 </script>
+
+<svelte:window use:action_exit={{on_bind: on_dismiss}} />
 
 {#if $_overlay_id}
     <input
@@ -148,7 +152,14 @@
         "spacing-x": spacing_x,
         "spacing-y": spacing_y,
     })}
-    on:click={on_click_inside}
+    use:click_inside={{
+        ignore: `label[for="${$_overlay_id}"]`,
+        on_click_inside: on_once,
+    }}
+    use:click_outside={{
+        ignore: `label[for="${$_overlay_id}"]`,
+        on_click_outside: on_dismiss,
+    }}
     use:trap_focus={{enabled: $_overlay_state, first: focus_first, last: focus_last}}
     use:auto_focus={{enabled: $_overlay_state, target: focus_target}}
     use:forward_actions={{actions}}